LibWeb: Split ContentStyleValue out of StyleValue.{h,cpp}
